Support global uninitialized pthread_mutex_t instances.
If such an instance is placed in a global/static section, the complier will initialize it with zeroes and pthread
code will perform implicit initialization of such a mutex upon its first usage. This is used by some Unix apps
that omit initializing such an instance with PTHREAD_MUTEX_INITIALIZER violating Posix requirement.

Supporting this violating behavior on OS/2 is rather safe and is definitely simpler than fixing each and every
incorrect usage.

Closes #8.