Guard against no valid pointer, part 2.
This adds a more strict check via a magic to make sure that the right version
of pthread_condattr_init was called on attr.

It is still possible that somehow a random piece of memory passed in as attr
to pthread_cond_init contains the right magic at the right location but this
is highly unlikely.

Closes #16.